test: split up EcsService tests
