+++ package.json
@@ -28,7 +28,7 @@
     "build": "npm run clean && tsc",
     "format": "prettier -w .",
     "release": "npm run build && release-it",
-    "test": "TS_NODE_PROJECT=./tsconfig.test.json node --test --test-concurrency=none -r ts-node/register ./tests/**/*.test.ts"
+    "test": "TS_NODE_PROJECT=./tsconfig.test.json node --test --test-concurrency=none -r ts-node/register ./tests/**/index.test.ts"
   },
   "prettier": "@studion/prettier-config",
   "dependencies": {
